Update UVLFs.py
This section of the code is made to apply dust recursively to the UVLF until convergence is reached. The bug made it such that dust would only be applied once (since currMUV2 would be equal to currMUV by definition). This branch fixes the bug and ensures the intended behavior of this process.